Updated frontend framework. fix package.json issue
We were using an incompatable version of typescript with Ionic, so I replaced that. npm install will be required

Set up the first route to home, created Home.tsx in pages and a sample Test.tsx in components.

Created a file ionicCheat.md that is a cheatsheet to help with building ionic pages. cheatsheet in not complete.